Updated Java and Selenium Versions
The current Selenium version is 3.141.59 (as of 2020-04/16), and the current legacy LTS Java version is 1.8. By now *everyone* is running either Java 1.8 or newer since 1.6 and 1.7 have been EOL'd for a while already. It should be safe to move up and those who can't, can remain with the old(er) Selenium and plugin binaries without issue.